requirements.txt
pandas
requests
stix2
PyYAML
